Allow user to specify wich comments export.py removes.
This is in preparation for modifying the Dual Extrusion merge logic so that it does not override the user's preferences with its own fixed defaults. The user must be able to specify (at least) that the <layer> and </layer> comments used by the Dual extrusion merge are not removed, even if other comments should be deleted.